requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
flask-restx==1.3.0
identify==2.6.3
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
queryparams-liberdade==0.0.1
referencing==0.35.1
rpds-py==0.22.3
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.28.0
Werkzeug==3.1.3
